requirements.txt
python-dotenv==0.21.1
python-telegram-bot==13.15
google-cloud-dialogflow==2.20.0
